Finalized initial Library class
Added method loadDecksFromFolder that allows Library to load
all decks from the local folder, so that Library can initialize
itself. Added necessary changes to other classes so that importing
and exporting work as planned. Fixed method removeDeck so it removes
not only deck entries in the vector, but also in the folder.